+++ src/app/helpers/pixi-app-setup.ts
@@ -1,11 +1,5 @@
 import { Application, Container } from 'pixi.js';
-
-export interface PixiAppConfig {
-  width: number;
-  height: number;
-  backgroundAlpha?: number;
-  antialias?: boolean;
-}
+import { PixiAppConfig } from '../interfaces/pixi-config';
 
 /**
  * Initializes a pixijs with the specified configuration
