+++ src/app/interfaces/pixi-config.ts
@@ -1,0 +1,6 @@
+export interface PixiAppConfig {
+  width: number;
+  height: number;
+  backgroundAlpha?: number;
+  antialias?: boolean;
+}
