+++ src/app/interfaces/sprite.ts
@@ -1,0 +1,11 @@
+import { Sprite } from 'pixi.js';
+
+export interface NodeSpriteData {
+  terrain: Sprite;
+  object?: Sprite;
+  claimIndicator?: Sprite;
+}
+
+export interface NodeSprites {
+  [nodeKey: string]: NodeSpriteData;
+}
